Modified: -구매페이지 레이아웃 수정 -구매페이지 컴포넌트 분
