+++ src/pages/Payment/Payment.scss
@@ -1,174 +1,7 @@
-main {
+.payment {
   display: flex;
   flex-direction: column;
   align-items: center;
   margin-bottom: 100px;
-
-  table {
-    width: 100%;
-    height: 100px;
-    margin: auto;
-    font-size: 10px;
-    border-collapse: collapse;
-  }
-
-  table,
-  td,
-  th {
-    margin: 0px;
-    padding: 10px;
-    background-color: #ffffff;
-    border: 1px solid #e9e9e9;
-  }
-
-  .orderProductList {
-    display: inline-block;
-    width: 90%;
-    margin-bottom: 30px;
-    padding-bottom: 20px;
-    border-bottom: 1px solid black;
-
-    ul {
-      padding: 10px;
-      background-color: #fef8f3;
-      border-top: 1px solid #e7e4e3;
-      border-bottom: 1px solid #e7e4e3;
-
-      li {
-        color: #fa726f;
-        font-size: 11px;
-        text-align: left;
-
-        .fa-exclamation {
-          margin-right: 10px;
-          padding: 2px 5px 2px 5px;
-          border: 1px solid #e3e2e5;
-          background-color: white;
-        }
-      }
-    }
-
-    .tableTitle {
-      position: relative;
-      padding: 5px;
-      background-color: #f6f6f6;
-      border: 1px solid #fafafa;
-      text-align: center;
-
-      h3 {
-        display: inline-block;
-        font-size: 12px;
-      }
-    }
-
-    .orderDetailsTable {
-      padding: 10px;
-      background-color: #ffffff;
-
-      .tableStyle {
-        background-color: #fafafa;
-      }
-
-      .tableline {
-        border-right: 1px solid white;
-      }
-    }
-
-    .totalPrice {
-      background-color: #fafafa;
-      text-align: right;
-
-      span {
-        color: #008bcc;
-      }
-    }
-  }
-
-  h3 {
-    font-size: 12px;
-  }
-
-  .text {
-    margin-right: 10px;
-    font-size: 10px;
-    text-align: right;
-  }
-
-  .orderArea {
-    width: 90%;
-    margin-bottom: 40px;
-    padding: 10px 0px;
-
-    .deliveryInfoTable {
-      width: 100%;
-      padding: 10px 0px;
-      background-color: #ffffff;
-      border: 1px solid #e3e2e5;
-
-      .table {
-        display: grid;
-        grid-template-columns: 150px auto;
-        column-gap: 0px;
-
-        .deliveryInfo {
-          background-color: white;
-          margin-left: 20px;
-          text-align: left;
-
-          span {
-            font-size: 12px;
-          }
-        }
-
-        input {
-          width: 60%;
-          border: 1px solid #d5d5d5;
-          text-align: left;
-        }
-
-        .emailInfo {
-          padding-top: 10px;
-          font-size: 12px;
-          text-align: left;
-        }
-      }
-    }
-  }
-
-  .totalPayAmountTable {
-    width: 90%;
-    margin: auto;
-    margin-bottom: 10px;
-    border: 1px solid black;
-
-    .tableStyle {
-      background-color: #fafafa;
-    }
-
-    td {
-      text-align: center;
-    }
-  }
-
-  .payArea {
-    width: 90%;
-    padding: 5px;
-    background-color: white;
-    text-align: right;
-    margin-right: 10px;
-
-    p {
-      color: #008bcc;
-      margin-bottom: 10px;
-    }
-
-    button {
-      width: 200px;
-      border: none;
-      color: white;
-      background-color: #3a4050;
-      border-radius: 5px;
-      padding: 10px;
-    }
-  }
+  text-align: center;
 }
