+++ src/pages/Payment/PaymentOrderDetails.scss
@@ -1,0 +1,75 @@
+.paymentOrderDetails {
+  width: 90%;
+  margin-bottom: 30px;
+  padding-bottom: 50px;
+  border-bottom: 1px solid black;
+
+  table {
+    width: 100%;
+    height: 100px;
+    margin: auto;
+    border-collapse: collapse;
+    font-size: 10px;
+  }
+
+  td,
+  th {
+    padding: 10px;
+    background-color: #ffffff;
+    border: 1px solid #e9e9e9;
+  }
+
+  .orderProductList {
+    ul {
+      padding: 10px;
+      background-color: #fef8f3;
+      border-top: 1px solid #e7e4e3;
+      border-bottom: 1px solid #e7e4e3;
+
+      li {
+        color: #fa726f;
+        font-size: 11px;
+        text-align: left;
+
+        .faExclamation {
+          margin-right: 10px;
+          padding: 2px 6px 2px 5px;
+          border: 1px solid #e3e2e5;
+          background-color: white;
+        }
+      }
+    }
+
+    .tableTitle {
+      padding: 5px;
+      background-color: #f6f6f6;
+      text-align: center;
+
+      h3 {
+        font-size: 12px;
+      }
+    }
+
+    .orderDetailsTable {
+      padding: 10px;
+      background-color: #ffffff;
+
+      .tableStyle {
+        background-color: #fafafa;
+      }
+
+      .tableline {
+        border-right: 1px solid white;
+      }
+    }
+
+    .totalPrice {
+      background-color: #fafafa;
+      text-align: right;
+
+      span {
+        color: #008bcc;
+      }
+    }
+  }
+}
